backend/requirements.txt
asgiref==3.2.3
astroid==2.4.0
autopep8==1.5
Babel==2.8.0
backcall==0.1.0
cassandra-driver==3.22.0
certifi==2019.11.28
chardet==3.0.4
click==7.1.1
decorator==4.4.1
Django==3.0.2
django-background-tasks==1.2.5
django-cassandra-engine==1.6.0
django-compat==1.0.15
django-cors-headers==3.4.0
django-crispy-forms==1.9.0
django-extensions==3.0.3
django-filter==2.2.0
django-phonenumber-field==4.0.0
django-rest==0.8.6
django-rest-registration==0.5.6
django-shell-plus==1.1.7
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
djangorestframework-oauth==1.1.0
geomet==0.1.2
httpie==2.0.0
idna==2.8
ipython==7.11.1
ipython-genutils==0.2.0
isort==4.3.21
jedi==0.16.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
parso==0.6.0
pexpect==4.8.0
phonenumbers==8.12.0
pickleshare==0.7.5
Pillow==7.0.0
prompt-toolkit==3.0.3
ptyprocess==0.6.0
pycodestyle==2.5.0
Pygments==2.5.2
PyJWT==1.7.1
pylint==2.5.0
pytz==2019.3
PyYAML==5.3.1
razorpay==1.2.0
requests==2.22.0
six==1.14.0
sqlparse==0.3.0
toml==0.10.0
traitlets==4.3.3
typed-ast==1.4.1
typing==3.7.4.1
uritemplate==3.0.1
urllib3==1.25.8
wcwidth==0.1.8
wrapt==1.12.1
